requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==20.6
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
yt-dlp==2025.3.21
